Computation logic moved to helper funtction in separate file
