+++ cockatrice/CMakeLists.txt
@@ -82,6 +82,7 @@ set(cockatrice_SOURCES
     src/game_graphics/game_scene.cpp
     src/game/game_state.cpp
     src/game_graphics/game_view.cpp
+    src/game/selection_subtype_counter.cpp
     src/game_graphics/hand_counter.cpp
     src/game_graphics/log/message_log_widget.cpp
     src/game/phase.cpp
