+++ cockatrice/src/game/selection_subtype_counter.cpp
@@ -1,0 +1,129 @@
+#include "selection_subtype_counter.h"
+
+#include "board/card_item.h"
+
+#include <QMap>
+#include <algorithm>
+
+namespace SelectionSubtypeCounter
+{
+
+QStringList extractSubtypesFromFace(const QString &faceType)
+{
+    QStringList parts = faceType.split(QStringLiteral(" — "));
+    if (parts.size() > 1) {
+        return parts[1].split(QStringLiteral(" "), Qt::SkipEmptyParts);
+    }
+    return {};
+}
+
+QList<MainTypeGroup> countSubtypes(const QList<CardItem *> &cards)
+{
+    QMap<QString, QMap<QString, int>> subtypesByMainType;
+    QMap<QString, int> cardCountPerMainType;
+
+    for (CardItem *card : cards) {
+        if (card->getFaceDown() || card->getCard().isEmpty()) {
+            continue;
+        }
+
+        QString mainType = card->getCardInfo().getMainCardType();
+        if (mainType.isEmpty()) {
+            mainType = QStringLiteral("Other");
+        }
+
+        QString cardType = card->getCardInfo().getCardType();
+        QStringList cardFaces = cardType.split(QStringLiteral(" // "));
+
+        bool contributedSubtypes = false;
+        for (const QString &face : cardFaces) {
+            QStringList subtypes = extractSubtypesFromFace(face);
+            for (const QString &subtype : subtypes) {
+                subtypesByMainType[mainType][subtype]++;
+                contributedSubtypes = true;
+            }
+        }
+
+        if (contributedSubtypes) {
+            cardCountPerMainType[mainType]++;
+        }
+    }
+
+    QList<MainTypeGroup> groups;
+    for (auto it = subtypesByMainType.constBegin(); it != subtypesByMainType.constEnd(); ++it) {
+        MainTypeGroup group;
+        group.mainType = it.key();
+        group.cardCount = cardCountPerMainType.value(it.key(), 0);
+
+        for (auto subIt = it.value().constBegin(); subIt != it.value().constEnd(); ++subIt) {
+            group.subtypes.append({subIt.key(), subIt.value()});
+        }
+
+        // Sort subtypes: by count ascending, then alphabetically
+        std::sort(group.subtypes.begin(), group.subtypes.end(), [](const SubtypeEntry &a, const SubtypeEntry &b) {
+            if (a.count != b.count) {
+                return a.count < b.count;
+            }
+            return a.name < b.name;
+        });
+
+        groups.append(group);
+    }
+
+    // Sort groups: by card count ascending, then alphabetically by main type
+    std::sort(groups.begin(), groups.end(), [](const MainTypeGroup &a, const MainTypeGroup &b) {
+        if (a.cardCount != b.cardCount) {
+            return a.cardCount < b.cardCount;
+        }
+        return a.mainType < b.mainType;
+    });
+
+    return groups;
+}
+
+QString formatAsHtml(const QList<MainTypeGroup> &groups)
+{
+    // Flatten to final ordered list
+    QList<SubtypeEntry> sortedEntries;
+    for (const MainTypeGroup &group : groups) {
+        for (const auto &entry : group.subtypes) {
+            sortedEntries.append(entry);
+        }
+    }
+
+    // Calculate padding widths
+    int maxNameLen = 0;
+    int maxCountLen = 0;
+    for (const auto &entry : sortedEntries) {
+        maxNameLen = qMax(maxNameLen, entry.name.length());
+        maxCountLen = qMax(maxCountLen, QString::number(entry.count).length());
+    }
+
+    // Format output
+    QStringList lines;
+    for (const auto &entry : sortedEntries) {
+        QString name = entry.name.toHtmlEscaped();
+        QString count = QString::number(entry.count);
+
+        QString namePadding = QString(QStringLiteral("&nbsp;")).repeated(maxNameLen - entry.name.length());
+        QString countPadding = QString(QStringLiteral("&nbsp;")).repeated(maxCountLen - count.length());
+
+        lines << QStringLiteral(
+                     "%1%2  <span style='font-size:14px;font-weight:bold;vertical-align:middle;'>%3%4</span>")
+                     .arg(namePadding, name, countPadding, count);
+    }
+
+    return QStringLiteral("<span style='font-family: monospace;'>") + lines.join(QStringLiteral("<br>")) +
+           QStringLiteral("</span>");
+}
+
+QString buildSubtypeCountText(const QList<CardItem *> &cards)
+{
+    QList<MainTypeGroup> groups = countSubtypes(cards);
+    if (groups.isEmpty()) {
+        return QString();
+    }
+    return formatAsHtml(groups);
+}
+
+} // namespace SelectionSubtypeCounter
